+++ src/main/java/com/example/demo/user/controller/UserController.java
@@ -2,27 +2,18 @@ package com.example.demo.user.controller;
 
 
 
-import com.example.demo.security.service.UserDetailsService;
 import com.example.demo.user.service.LoginRequest;
 import com.example.demo.user.service.SignupRequest;
 import com.example.demo.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import lombok.Value;
 
-import org.springframework.web.bind.annotation.*;
 
-import javax.servlet.Servlet;
+import org.springframework.web.bind.annotation.*;
 import javax.servlet.ServletException;
-import javax.servlet.ServletOutputStream;
-import javax.servlet.ServletResponse;
-import javax.servlet.http.Cookie;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
+
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.io.PrintWriter;
-import java.util.Collection;
-import java.util.Locale;
 
 @RestController
 @RequestMapping("/account")
@@ -56,8 +47,13 @@ public class UserController {
     }
 
     @PostMapping("/login")
-    public void Login(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        userService.login(request, response);
+    public void Login(@RequestBody LoginRequest loginRequest, HttpServletResponse response) throws ServletException, IOException {
+        String username = loginRequest.getUsername();
+        String password = loginRequest.getPassword();
+
+//        UsernamePasswordAuthenticationToken token
+//                = new UsernamePasswordAuthenticationToken(username, password);
+//        SecurityContextHolder.ge
     }
 
     @Value
