+++ src/main/java/com/example/demo/user/service/UserService.java
@@ -4,18 +4,14 @@ package com.example.demo.user.service;
 
 
 
-import com.example.demo.security.infrastructure.UserAuthenticationFilter;
-import com.example.demo.security.service.UserDetailsService;
 import com.example.demo.user.domain.*;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.servlet.FilterChain;
 import javax.servlet.ServletException;
-import javax.servlet.ServletRequest;
-import javax.servlet.ServletResponse;
+
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
